requirements.txt
pandas
mysql-connector-python
faker
